requirements.txt
py-getch==1.0.1
requests==2.25.1
clint==0.5.1
